implement `CheckDiskUsage()` for Windows
